Initial commit for the prerequisites-tester.
This commit includes a basic test framework infrastructure, focusing on convenient logging, http-requests construction and response handling, and minimal test-management support.

The commit also includes the first two tests - one for testing the long polling mechanism between the agent and the server, and one for testing the websocket connection establishment and maintenance using either pinned or root certs, and a ping-pong mechanism to monitor the connection.
The ping-pong mechanism is still a work in progress, and will be fixed in future commits.
Basic documentation should also be included in a later commit.